Improve MapSheets printing file size
Now we cut the geometries by the view extent
before printing them in order to reduce
the resulting file size (gvSIG did print all
geometries that were returned by the driver).

Also tried to fix what seems to be an old bug
in MapSheets that we hadn't located, which
results in some of the printed maps not being
precisely adjusted to the grid. This proved
even more troublesome now that we only print
the required geometries. Couldn't find the
exact cause, but a workaround has been applied
to fix the problem.

refs #1110