Two field on enlaces are automatically calculated
"numero_ramales" and "longitud_ramales" are automatically calculated
with a db trigger on the tables "ramales" and "enlaces". To "notify"
the "Enlaces" layer and the "Enlaces" form about this change custom
code is writen in `RamalesForm.saveRecord`.

Still, if a "ramal" is removed, this logic, fails. Probably
`AbstractNavtable.deleteRecord` should be also overriden

refs #1779