ELLE refactor and new feature to export/load simbology and labels
Two elle classes that handles working with simbology and labels are
moved to commons.

Also ExportTo is modified. When a FLyVect layer of single type geoms
is exported to shape a .gvl and a .gvlabel files are also created with
the xml definition of the simbology and the label.

The VectorialFileOpen class (called when the user uses AddLayer) is modified
to load and apply the .gvlabel files when loading a shape

refs #1684