Adding 'trabajos' in batch mode is slow
When there are many elements with 'trabajos' to be added in batch form
the table goes to a open edition, close edition (commit to db) for each
job. This commit fixes this.

* A new method in IController allows to call create in batch mode so the
  caller has the responsability to open/close the edition
* BatchAbstractSubForm makes use of the new method

Also some refactoring is done, and the iteration through selected elements
on the main layer is optimized. We just go to the next selected element instead
of iterate all elements asking if it is selected

refs #2154