Updates postgis and postgresql jdbc drivers
Libraries

* libGDBMS/lib/postgresql-8.0-310.jdbc3.jar. I'm not sure about this jar. It
  seems that is not being used. We just remove it.
* libFMap/lib/postgis-jdbc-driver.jar. Old postgis driver. We remove it.
* libFMap/lib/postgresql-74.213.jar. Old postgresql driver. We remote it.
* extGIA/lib/postgresql-9.2-1003.jdbc3.jar. A more up-to-date library used in
  own code in a workaround to fix issues with working with blobs and ELLE.
* libFMap/lib/postgis-jdbc-2.1.7.2.jar. New postgis jdbc library.
* libFMap/lib/postgresql-42.2.27.jre6.jar. New postgresql jdbc library.

Classpath and compilation issues

* libIverUtiles were depending on postgis and postgresql driver, but after
  remove those dependencies from the classpath it compiles. So it does not
  depend on the new libs
* libGBDMS depends on old jar, and the new ones should be added to the
  classpath to avoid compilation issues. Anyway only one class in this
  package depends on them. Probably it should be moved to other place
* extGIA depends on old postgresql driver but it's fine just remove it

extJDCB

Depends initially on both drivers. But after find nio exceptions working with
the application and refactoring a bit the code on compilation it only depends
of postgis. But, when compiling the full workspace, is needed to add postgresql
also.

A broken unit test that does not compile is also removed

The changes in this packages should be tested carefully

libFMap

"next" can not be called on the ResultSet after calling "moveToInsert", it's
removed. Anyway, the cursor is always updated before update/delete/insert a
row so that "next" it's not needed

extGIA

No need to use our own driver here to handle images/blobs, so related stuff
is removed